create handle_client module
